Create Jackpot
This program decides how much to pay contestants from jackpot contestants based on the answers of an automated survey. There are exactly 6 contestants and 6 question. There are two additional features that must be added 1) Pay out money to people's bank account using the internet. 2) Agencies in the network torus are awarded based on a lottery where there probability of success based on how well their agency does compared to the total. Then we must make the retrieval from the SQLite database secure.