Create Surveyquestion
A lot needs to be changed. 1) Self eval should be rewritten as peer eval 2) People listed should be in a random order 3) In order to run the second part we have to prematurely exit the first window, this is bad 4) The list used needs to change as not to over index 5) Center tkinter window in the middle 5) Make tkinter window robust when it comes to password type stuff 6) I am incredibly extroverted, so not working on this anymore until others join. Feel free to email me at numbercomputing@gmail.com with the title "Microfinance Refugees".